requirements.txt
pyqt5-tools
flask
pyusb
pynput